Add version to server interactions
It is very easy for the clients' state to become out of sync with the
MZC server state. We add a version to the GenServer state, and if the
passed version doesn't match the current state version, the current
state is sent back so that the client can redisplay the new version.